10_LLM-service/02_langchain/01_lecture-source/04_simple-chatbot.py/03_chatbot-with-vector-store/requirements.txt
fastapi
uvicorn
gunicorn
langchain
langchain-community
langchain-openai
python-dotenv
python-multipart
faiss-cpu